requirements.txt
numpy
pandas
matplotlib
opencv-python
tensorflow
keras
scikit-learn
flask
pillow
python-docx
pypdf
python-dotenv
pytesseract
